+++ workflow-gatk-haplotype-caller/src/main/java/ca/on/oicr/pde/workflows/GATKHaplotypeCallerWorkflow.java
@@ -28,18 +28,10 @@
 package ca.on.oicr.pde.workflows;
 
 import ca.on.oicr.pde.tools.BEDFileUtils;
-import ca.on.oicr.pde.tools.gatk3.AnalyzeCovariates;
-import ca.on.oicr.pde.tools.gatk3.BaseRecalibrator;
 import ca.on.oicr.pde.tools.gatk3.CatVariants;
 import ca.on.oicr.pde.tools.gatk3.HaplotypeCaller;
-import ca.on.oicr.pde.tools.gatk3.IndelRealigner;
-import ca.on.oicr.pde.tools.gatk3.PrintReads;
-import ca.on.oicr.pde.tools.gatk3.RealignerTargetCreator;
 import ca.on.oicr.pde.utilities.workflows.OicrWorkflow;
-import com.google.common.collect.HashMultimap;
 import com.google.common.collect.Iterables;
-import com.google.common.collect.Multimap;
-import com.google.common.collect.Sets;
 import java.io.IOException;
 import java.util.*;
 import java.util.Map.Entry;
@@ -139,7 +131,6 @@ public class GATKHaplotypeCallerWorkflow extends OicrWorkflow {
         final String queue = getOptionalProperty("queue", "");
         final String java = getProperty("java");
         final String gatk = getOptionalProperty("gatk_jar", binDir);
-        final String rDir = getProperty("r_dir");
         final String gatkKey = getProperty("gatk_key");
         final String identifier = getProperty("identifier");
         final String refFasta = getProperty("ref_fasta");
@@ -147,25 +138,10 @@ public class GATKHaplotypeCallerWorkflow extends OicrWorkflow {
         final Integer intervalPadding = hasPropertyAndNotNull("interval_padding") ? Integer.parseInt(getProperty("interval_padding")) : null;
         final Integer downsamplingCoverage = hasPropertyAndNotNull("downsampling_coverage") ? Integer.parseInt(getProperty("downsampling_coverage")) : null;
         final String downsamplingType = getOptionalProperty("downsampling_type", null);
-        final Integer preserveQscoresLessThan = hasPropertyAndNotNull("preserve_qscores_less_than") ? Integer.parseInt(getProperty("preserve_qscores_less_than")) : null;
-        final Set<String> bqsrCovariates = Sets.newHashSet(StringUtils.split(getProperty("bqsr_covariates"), ","));
-        final Boolean doBQSR = Boolean.valueOf(getOptionalProperty("do_bqsr", "true"));
-        final Integer gatkRealignTargetCreatorXmx = Integer.parseInt(getProperty("gatk_realign_target_creator_xmx"));
-        final Integer gatkIndelRealignerXmx = Integer.parseInt(getProperty("gatk_indel_realigner_xmx"));
-        final Integer gatkPrintReadsXmx = Integer.parseInt(getProperty("gatk_print_reads_xmx"));
-        final Integer gatkBaseRecalibratorXmx = Integer.parseInt(getProperty("gatk_base_recalibrator_xmx"));
-        final Integer gatkBaseRecalibratorMem = Integer.parseInt(getProperty("gatk_base_recalibrator_mem"));
-        final Integer gatkBaseRecalibratorNct = Integer.parseInt(getProperty("gatk_base_recalibrator_nct"));
-        final Integer gatkBaseRecalibratorSmp = Integer.parseInt(getProperty("gatk_base_recalibrator_smp"));
         final Integer gatkHaplotypeCallerThreads = Integer.parseInt(getProperty("gatk_haplotype_caller_threads"));
         final Integer gatkHaplotypeCallerXmx = Integer.parseInt(getProperty("gatk_haplotype_caller_xmx"));
         final Integer gatkCombineGVCFsXmx = Integer.parseInt(getProperty("gatk_combine_gvcfs_xmx"));
         final Integer gatkOverhead = Integer.parseInt(getProperty("gatk_sched_overhead_mem"));
-        final String realignerTargetCreatorParams = getOptionalProperty("gatk_realigner_target_creator_params", null);
-        final String indelRealignerParams = getOptionalProperty("gatk_indel_realigner_params", null);
-        final String baseRecalibratorParams = getOptionalProperty("gatk_base_recalibrator_params", null);
-        final String analyzeCovariatesParams = getOptionalProperty("gatk_analyze_covariates_params", null);
-        final String printReadsParams = getOptionalProperty("gatk_print_reads_params", null);
         final String haplotypeCallerParams = getOptionalProperty("gatk_haplotype_caller_params", null);
 
         final List<String> intervalFilesList = Arrays.asList(StringUtils.split(getOptionalProperty("interval_files", ""), ","));
@@ -199,123 +175,11 @@ public class GATKHaplotypeCallerWorkflow extends OicrWorkflow {
             chrSizes.add(null);
         }
 
-        Multimap<String, Pair<String, Job>> realignedBams = HashMultimap.create();
-        for (String chrSize : chrSizes) {
-
-            //GATK Realigner Target Creator ( https://www.broadinstitute.org/gatk/gatkdocs/org_broadinstitute_gatk_tools_walkers_indels_RealignerTargetCreator.php )
-            RealignerTargetCreator realignerTargetCreatorCommand = new RealignerTargetCreator.Builder(java, gatkRealignTargetCreatorXmx + "g", tmpDir, gatk, gatkKey, dataDir)
-                    .setReferenceSequence(refFasta)
-                    .addInputBamFiles(inputBamFiles)
-                    .setKnownIndels(dbsnpVcf)
-                    .addInterval(chrSize)
-                    .addIntervalFiles(intervalFiles)
-                    .setIntervalPadding(intervalPadding)
-                    .setDownsamplingCoverageThreshold(downsamplingCoverage)
-                    .setDownsamplingType(downsamplingType)
-                    .setOutputFileName("gatk" + (chrSize != null ? "." + chrSize.replace(":", "-") : ""))
-                    .setExtraParameters(realignerTargetCreatorParams)
-                    .build();
-            Job realignerTargetCreatorJob = getWorkflow().createBashJob("GATKRealignerTargetCreator")
-                    .setMaxMemory(Integer.toString((gatkRealignTargetCreatorXmx + gatkOverhead) * 1024))
-                    .setQueue(queue);
-            realignerTargetCreatorJob.getCommand().setArguments(realignerTargetCreatorCommand.getCommand());
-
-            //GATK Indel Realigner ( https://www.broadinstitute.org/gatk/gatkdocs/org_broadinstitute_gatk_tools_walkers_indels_IndelRealigner.php )
-            IndelRealigner indelRealignerCommand = new IndelRealigner.Builder(java, gatkIndelRealignerXmx + "g", tmpDir, gatk, gatkKey, dataDir)
-                    .setReferenceSequence(refFasta)
-                    .addInputBamFiles(inputBamFiles)
-                    .addKnownIndelFile(dbsnpVcf)
-                    .addInterval(chrSize)
-                    .setIntervalPadding(intervalPadding)
-                    .setTargetIntervalFile(realignerTargetCreatorCommand.getOutputFile())
-                    .setExtraParameters(indelRealignerParams)
-                    .build();
-            Job indelRealignerJob = getWorkflow().createBashJob("GATKIndelRealigner")
-                    .setMaxMemory(Integer.toString((gatkIndelRealignerXmx + gatkOverhead) * 1024))
-                    .setQueue(queue)
-                    .addParent(realignerTargetCreatorJob);
-            indelRealignerJob.getCommand().setArguments(indelRealignerCommand.getCommand());
-
-            if (realignedBams.containsKey(chrSize)) {
-                throw new RuntimeException("Unexpected state: Duplicate interval key.");
-            }
-            for (String outputFile : indelRealignerCommand.getOutputFiles()) {
-                realignedBams.put(chrSize, Pair.of(outputFile, indelRealignerJob));
-            }
-        }
-
-        Multimap<String, Pair<String, Job>> inputBams;
-        if (doBQSR) {
-            Multimap<String, Pair<String, Job>> recalibratedBams = HashMultimap.create();
-
-            //GATK Base Recalibrator ( https://www.broadinstitute.org/gatk/gatkdocs/org_broadinstitute_gatk_tools_walkers_bqsr_BaseRecalibrator.php )
-            BaseRecalibrator baseRecalibratorCommand = new BaseRecalibrator.Builder(java, gatkBaseRecalibratorXmx + "m", tmpDir, gatk, gatkKey, dataDir)
-                    .setReferenceSequence(refFasta)
-                    .setCovariates(bqsrCovariates)
-                    .addKnownSite(dbsnpVcf)
-                    .addInputFiles(getLeftCollection(realignedBams.values()))
-                    .addIntervalFiles(intervalFiles)
-                    .setIntervalPadding(intervalPadding)
-                    .setNumCpuThreadsPerDataThread(gatkBaseRecalibratorNct)
-                    .setExtraParameters(baseRecalibratorParams)
-                    .build();
-            Job baseRecalibratorJob = getWorkflow().createBashJob("GATKBaseRecalibrator")
-                    .setMaxMemory(gatkBaseRecalibratorMem.toString())
-                    .setThreads(gatkBaseRecalibratorSmp)
-                    .setQueue(queue);
-            baseRecalibratorJob.getParents().addAll(getRightCollection(realignedBams.values()));
-            baseRecalibratorJob.getCommand().setArguments(baseRecalibratorCommand.getCommand());
-
-            //GATK Analyze Covariates ( https://www.broadinstitute.org/gatk/guide/tooldocs/org_broadinstitute_gatk_tools_walkers_bqsr_AnalyzeCovariates.php )
-            AnalyzeCovariates analyzeCovariatesCommand = new AnalyzeCovariates.Builder(java, "4g", tmpDir, gatk, gatkKey, rDir, dataDir)
-                    .setReferenceSequence(refFasta)
-                    .setRecalibrationTable(baseRecalibratorCommand.getOutputFile())
-                    .setOutputFileName(identifier)
-                    .setExtraParameters(analyzeCovariatesParams)
-                    .build();
-            Job analyzeCovariatesJob = getWorkflow().createBashJob("GATKAnalyzeCovariates")
-                    .setMaxMemory(Integer.toString((4 + gatkOverhead) * 1024))
-                    .setQueue(queue)
-                    .addParent(baseRecalibratorJob);
-            analyzeCovariatesJob.getCommand().setArguments(analyzeCovariatesCommand.getCommand());
-            analyzeCovariatesJob.addFile(createOutputFile(analyzeCovariatesCommand.getPlotsReportFile(), "application/pdf", manualOutput));
-
-            for (Entry<String, Pair<String, Job>> e : realignedBams.entries()) {
-
-                String chrSize = e.getKey();
-                String inputBam = e.getValue().getLeft();
-
-                //GATK Print Reads ( https://www.broadinstitute.org/gatk/gatkdocs/org_broadinstitute_gatk_tools_walkers_readutils_PrintReads.php )
-                PrintReads printReadsCommand = new PrintReads.Builder(java, gatkPrintReadsXmx + "g", tmpDir, gatk, gatkKey, dataDir)
-                        .setReferenceSequence(refFasta)
-                        .setCovariatesTablesFile(baseRecalibratorCommand.getOutputFile())
-                        .setInputFile(inputBam)
-                        .setPreserveQscoresLessThan(preserveQscoresLessThan)
-                        .addInterval(chrSize)
-                        .setIntervalPadding(intervalPadding)
-                        .setExtraParameters(printReadsParams)
-                        .build();
-                Job printReadsJob = getWorkflow().createBashJob("GATKTableRecalibration")
-                        .setMaxMemory(Integer.toString((gatkPrintReadsXmx + gatkOverhead) * 1024))
-                        .setQueue(queue)
-                        .addParent(baseRecalibratorJob);
-                printReadsJob.getCommand().setArguments(printReadsCommand.getCommand());
-
-                recalibratedBams.put(chrSize, Pair.of(printReadsCommand.getOutputFile(), printReadsJob));
-            }
-
-            //BQSR enabled, pass recalibrated bams to variant calling
-            inputBams = recalibratedBams;
-        } else {
-            //BQSR disabled, pass realigned bams to variant calling
-            inputBams = realignedBams;
-        }
-
         Map<String, Pair<HaplotypeCaller, Job>> gvcfs = new HashMap<>();
         for (String chrSize : chrSizes) {
             //GATK Haplotype Caller ( https://www.broadinstitute.org/gatk/gatkdocs/org_broadinstitute_gatk_tools_walkers_haplotypecaller_HaplotypeCaller.php )
             HaplotypeCaller haplotypeCallerCommand = new HaplotypeCaller.Builder(java, Integer.toString(gatkHaplotypeCallerXmx) + "g", tmpDir, gatk, gatkKey, dataDir)
-                    .setInputBamFiles(getLeftCollection(inputBams.values()))
+                    .setInputBamFiles(inputBamFiles)
                     .setReferenceSequence(refFasta)
                     .setDbsnpFilePath(dbsnpVcf)
                     .addInterval(chrSize)
@@ -330,7 +194,6 @@ public class GATKHaplotypeCallerWorkflow extends OicrWorkflow {
             Job haplotypeCallerJob = this.getWorkflow().createBashJob("GATKHaplotypeCaller")
                     .setMaxMemory(Integer.toString((gatkHaplotypeCallerXmx + gatkOverhead) * 1024))
                     .setQueue(queue);
-            haplotypeCallerJob.getParents().addAll(getRightCollection(inputBams.values()));
             haplotypeCallerJob.getCommand().setArguments(haplotypeCallerCommand.getCommand());
 
             if (gvcfs.put(chrSize, Pair.of(haplotypeCallerCommand, haplotypeCallerJob)) != null) {